all: gcc warning clean round 4 * remove unused variable * move static function from h into cpp * Initialized hw_by_page to 0xFFFFFFFF instead of -1 (number must be a positive integer) * Use a s32 fore m_current_lsn instead of u32 (use -1 as error code) Bonus: a couple of fix for clang compiler (doesn't mean that it fully compile with clang) * remove useless __debugbreak on linux * use short for 16bits atomic function
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@5695 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ pcsx2/IPU/mpeg2lib/Idct.cpp
@@ -213,7 +213,7 @@ __ri void mpeg2_idct_add (const int last, s16 * block, s16 * dest, const int str
     }
     else
     {
-		s16 DC = (block[0] + 4) >> 3;
+		s16 DC = ((int)block[0] + 4) >> 3;
 		s16 dcf[2] = { DC, DC };
 		block[0] = block[63] = 0;
